fix: 로그아웃 API 스펙 변경
1. 로그아웃 API 스펙 변경
  - Authroization 토큰과 refreshToken을 같이 검증
    - refreshToken의 recordKey와 Authorization의 recordKey가 일치할 경우 로그아웃 진행

+++ src/main/java/com/github/sunguk0810/assignment/domain/auth/service/AuthServiceImpl.java
@@ -23,6 +23,7 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
+import java.util.Optional;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -65,7 +66,6 @@ public class AuthServiceImpl implements AuthService{
     @Override
     public TokenResponse refresh(RefreshTokenRequest request) {
         String refreshToken = request.getRefreshToken();
-        log.info("refreshToken = {}", refreshToken);
         TokenInfo tokenInfo = tokenRepository
                 .findByRefreshToken(refreshToken)
                 .orElseThrow(() -> new BusinessException(ErrorType.REFRESH_TOKEN_NOT_FOUND));
@@ -95,13 +95,14 @@ public class AuthServiceImpl implements AuthService{
     }
 
     @Override
-    public Boolean logout(LogoutRequest request) {
-        TokenInfo tokenInfo = tokenRepository
-                .findByRefreshToken(request.getRefreshToken())
+    public Boolean logout(String recordKey, LogoutRequest request) {
+        TokenInfo tokenInfo = tokenRepository.findByRefreshToken(request.getRefreshToken())
                 .orElseThrow(() -> new BusinessException(ErrorType.REFRESH_TOKEN_NOT_FOUND));
 
+        if (!tokenInfo.getRecordKey().equals(recordKey)){
+            throw new BusinessException(ErrorType.UNAUTHORIZED);
+        }
         tokenRepository.delete(tokenInfo);
-
         return true;
     }
 }
